% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_plotter.R
\name{matrix_plotter}
\alias{matrix_plotter}
\title{Different default methods to visualize matrices}
\usage{
matrix_plotter(data, method, col = NULL, coef = TRUE, is.corr = TRUE)
}
\arguments{
\item{data}{The matrix to be visualized.}

\item{method}{Current options: 'corrplot', 'heatmap'.}

\item{col}{(default = NULL) Colors for the values to be plotted.}

\item{coef}{(default = TRUE) Whether to display the values of the matrix}

\item{is.corr}{(default = TRUE) Whether the matrix is a correlation matrix}
}
\description{
Different default methods to visualize matrices
}
