% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mora_bada.R
\name{mora_bada}
\alias{mora_bada}
\title{BADA with inference and graphs.}
\usage{
mora_bada(
  data,
  design,
  col4obs,
  col4group,
  center = TRUE,
  scale = "SS1",
  make_design_nominal = TRUE,
  test.iters = 1000,
  important = FALSE
)
}
\arguments{
\item{data}{Numerical data table with observations on rows and variables on columns.}

\item{design}{A vector or dummy-coded matrix indicating group membership of rows.}

\item{col4obs}{A single color or vector of colors whose length is equal to nrow(data).}

\item{col4group}{A single color or vector of colors whose length is the number of groups in design.}

\item{center}{(default = TRUE) Whether to center variables}

\item{scale}{(default = "SS1") Whether to scale variables}

\item{make_design_nominal}{(default = TRUE) If design is a dummy-coded matrix, should be set to FALSE.}

\item{test.iters}{(default = 1000) Number of iterations for bootstrap.}

\item{important}{(default = FALSE) If TRUE, graphs have only the important saliences/bootstrap ratios.}
}
\description{
BADA with inference and graphs.
}
