% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mora_pca.R
\name{mora_pca}
\alias{mora_pca}
\title{PCA with inference and graphs.}
\usage{
mora_pca(
  data,
  design = NULL,
  make_design_nominal = TRUE,
  col4obs = "olivedrab3",
  col4group = "olivedrab3",
  center = TRUE,
  scale = "SS1",
  want34 = FALSE,
  inference = TRUE
)
}
\arguments{
\item{data}{The numerical data for PCA.}

\item{design}{(default = NULL) A vector or dummy-coded matrix that describes the rows.}

\item{make_design_nominal}{(default = TRUE) If design is a dummy-coded matrix, should be set to FALSE.}

\item{col4obs}{(default = "olivedrab3") A single color or vector of colors whose length is equal to nrow(data).}

\item{col4group}{(default = "olivedrab3") A single color or vector of colors whose length is the number of groups in design.}

\item{center}{(default = TRUE) Whether to center variables}

\item{scale}{(default = "SS1") Whether to scale variables}

\item{want34}{(default = FALSE) By default, only prints dimensions 1 and 2. Set to TRUE for 3 and 4.}

\item{inference}{(default = TRUE) When design contains 2 or more groups, computes CI and TI on observations. FALSE if no groups.}
}
\description{
PCA with inference and graphs.
}
