% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mora_plsc.R
\name{mora_plsc}
\alias{mora_plsc}
\title{PLSC with inferences and graphs.}
\usage{
mora_plsc(
  data1,
  data2,
  design = NULL,
  make_design_nominal = TRUE,
  col4obs = "olivedrab3",
  col4group = "olivedrab3",
  center1 = TRUE,
  center2 = TRUE,
  scale1 = "SS1",
  scale2 = "SS1",
  inference = TRUE,
  important = FALSE
)
}
\arguments{
\item{data1}{A numerical data table with observations on rows and variables on columns.}

\item{data2}{A numerical data table with the same obs on row and different vars on columns.}

\item{design}{(default = NULL) A design vector or matrix for the rows}

\item{make_design_nominal}{(default = TRUE) If TRUE, design is a vector. If FALSE, design is a dummy-coded matrix.}

\item{col4obs}{(default = "olivedrab3") A single color or vector of colors whose length is equal to nrow(data1).}

\item{col4group}{(default = "olivedrab3") A single color or vector of colors whose length is the number of groups in design.}

\item{center1}{(default = TRUE) Whether to center variables in data1.}

\item{center2}{(default = TRUE) Whether to center variables in data2.}

\item{scale1}{(default = "SS1") Whether to scale variables in data1.}

\item{scale2}{(default = "SS1") Whether to scale variables in data2.}

\item{inference}{(default = TRUE) When design contains 2 or more groups, computes CI and TI on observations. FALSE if no groups.}

\item{important}{(default = FALSE) If TRUE, graphs have only the important saliences/bootstrap ratios.}
}
\description{
PLSC with inferences and graphs.
}
